requirements.txt

python-dotenv==1.0.0
imaplib2==3.06
requests==2.31.0
pandas==2.1.1
pydantic==2.4.2
openai==1.2.0